+++ src/builtins.h
@@ -27,8 +27,12 @@ int builtin_getopts(char **args);
 int builtin_eval(char **args);
 int builtin_printf(char **args);
 int builtin_echo(char **args);
+int builtin_history(char **args);
+int builtin_fc(char **args);
 int builtin_exec(char **args);
 int builtin_command(char **args);
+int builtin_time(char **args);
+int builtin_times(char **args);
 const char **get_builtin_names(void);
 const char *get_alias(const char *name);
 void load_aliases(void);
