+++ src/builtins_misc.c
@@ -26,45 +26,11 @@
 #include <sys/stat.h>
 #include <fnmatch.h>
 #include <stdint.h>
-#include <time.h>
-#include <sys/times.h>
 #include <sys/resource.h>
 
 extern int last_status;
 extern FILE *parse_input;
 #include <sys/wait.h>
-#include <signal.h>
-
-/* Map signal names to numbers for trap builtin. */
-static const struct { const char *n; int v; } sig_map[] = {
-    {"INT", SIGINT}, {"TERM", SIGTERM}, {"HUP", SIGHUP},
-#ifdef SIGQUIT
-    {"QUIT", SIGQUIT},
-#endif
-#ifdef SIGUSR1
-    {"USR1", SIGUSR1},
-#endif
-#ifdef SIGUSR2
-    {"USR2", SIGUSR2},
-#endif
-#ifdef SIGCHLD
-    {"CHLD", SIGCHLD},
-#endif
-#ifdef SIGCONT
-    {"CONT", SIGCONT},
-#endif
-#ifdef SIGSTOP
-    {"STOP", SIGSTOP},
-#endif
-#ifdef SIGALRM
-    {"ALRM", SIGALRM},
-#endif
-    {NULL, 0}
-};
-
-char *trap_cmds[NSIG];
-char *exit_trap_cmd;
-extern void trap_handler(int);
 extern void run_exit_trap(void);
 
 /* Exit the shell, freeing resources and using the provided status
@@ -165,139 +131,7 @@ int builtin_echo(char **args)
     return 1;
 }
 
-/* Display the command history or modify it with -c to clear or -d N to
- * delete a specific entry. */
-int builtin_history(char **args) {
-    if (args[1]) {
-        if (strcmp(args[1], "-c") == 0 && !args[2]) {
-            clear_history();
-            return 1;
-        } else if (strcmp(args[1], "-d") == 0 && args[2] && !args[3]) {
-            int id = atoi(args[2]);
-            delete_history_entry(id);
-            return 1;
-        } else {
-            fprintf(stderr, "usage: history [-c|-d NUMBER]\n");
-            return 1;
-        }
-    }
-    print_history();
-    return 1;
-}
 
-/* Replay or edit commands from history. */
-int builtin_fc(char **args) {
-    int list = 0;
-    const char *editor = NULL;
-    int i = 1;
-    for (; args[i] && args[i][0] == '-'; i++) {
-        if (strcmp(args[i], "-l") == 0) {
-            list = 1;
-        } else if (strcmp(args[i], "-e") == 0 && args[i+1]) {
-            editor = args[i+1];
-            i++;
-        } else {
-            fprintf(stderr, "usage: fc [-l] [-e editor] [first [last]]\n");
-            return 1;
-        }
-    }
-
-    int max_id = 1;
-    while (history_get_by_id(max_id))
-        max_id++;
-    max_id -= 1;
-    if (max_id <= 0)
-        return 1;
-
-    int first_id = max_id;
-    int last_id = max_id;
-    if (args[i]) {
-        first_id = atoi(args[i]);
-        if (first_id < 0)
-            first_id = max_id + first_id + 1;
-        if (args[i+1]) {
-            last_id = atoi(args[i+1]);
-            if (last_id < 0)
-                last_id = max_id + last_id + 1;
-        } else {
-            last_id = first_id;
-        }
-    }
-    if (first_id <= 0 || last_id <= 0 || first_id > max_id || last_id > max_id) {
-        fprintf(stderr, "fc: history range out of bounds\n");
-        return 1;
-    }
-    if (first_id > last_id) {
-        int tmp = first_id;
-        first_id = last_id;
-        last_id = tmp;
-    }
-
-    if (list) {
-        for (int id = first_id; id <= last_id; id++) {
-            const char *cmd = history_get_by_id(id);
-            if (cmd)
-                printf("%d %s\n", id, cmd);
-        }
-        return 1;
-    }
-
-    char tmpname[] = "/tmp/vush_fcXXXXXX";
-    int fd = mkstemp(tmpname);
-    if (fd < 0) {
-        perror("mkstemp");
-        return 1;
-    }
-    FILE *f = fdopen(fd, "w+");
-    if (!f) {
-        perror("fdopen");
-        close(fd);
-        unlink(tmpname);
-        return 1;
-    }
-    for (int id = first_id; id <= last_id; id++) {
-        const char *cmd = history_get_by_id(id);
-        if (cmd)
-            fprintf(f, "%s\n", cmd);
-    }
-    fflush(f);
-
-    if (!editor)
-        editor = getenv("FCEDIT");
-    if (!editor || !*editor)
-        editor = "ed";
-
-    pid_t pid = fork();
-    if (pid == 0) {
-        execlp(editor, editor, tmpname, NULL);
-        perror(editor);
-        _exit(127);
-    } else if (pid > 0) {
-        int status;
-        waitpid(pid, &status, 0);
-    } else {
-        perror("fork");
-        fclose(f);
-        unlink(tmpname);
-        return 1;
-    }
-
-    fseek(f, 0, SEEK_SET);
-    char line[MAX_LINE];
-    while (fgets(line, sizeof(line), f)) {
-        size_t len = strlen(line);
-        if (len && line[len-1] == '\n')
-            line[len-1] = '\0';
-        Command *cmds = parse_line(line);
-        if (cmds && cmds->pipeline && cmds->pipeline->argv[0])
-            run_command_list(cmds, line);
-        free_commands(cmds);
-    }
-
-    fclose(f);
-    unlink(tmpname);
-    return 1;
-}
 
 
 /* Manage or display command hash table. */
@@ -563,66 +397,7 @@ int builtin_command(char **args) {
 }
 
 
-/* Run a command and report the elapsed real time. */
-int builtin_time(char **args) {
-    if (!args[1]) {
-        fprintf(stderr, "usage: time command [args...]\n");
-        return 1;
-    }
 
-    struct timespec start, end;
-    clock_gettime(CLOCK_MONOTONIC, &start);
-    pid_t pid = fork();
-    if (pid == 0) {
-        execvp(args[1], &args[1]);
-        perror(args[1]);
-        _exit(127);
-    } else if (pid > 0) {
-        int status;
-        waitpid(pid, &status, 0);
-        clock_gettime(CLOCK_MONOTONIC, &end);
-        double elapsed = (end.tv_sec - start.tv_sec) +
-                         (end.tv_nsec - start.tv_nsec) / 1e9;
-        printf("real %.3f sec\n", elapsed);
-        if (WIFEXITED(status))
-            last_status = WEXITSTATUS(status);
-        else if (WIFSIGNALED(status))
-            last_status = 128 + WTERMSIG(status);
-        else
-            last_status = status;
-        return 1;
-    } else {
-        perror("fork");
-        last_status = 1;
-        return 1;
-    }
-}
-
-/* Print user/system CPU times for the shell and its children. */
-int builtin_times(char **args) {
-    if (args[1]) {
-        fprintf(stderr, "usage: times\n");
-        last_status = 1;
-        return 1;
-    }
-
-    struct tms t;
-    if (times(&t) == (clock_t)-1) {
-        perror("times");
-        last_status = 1;
-        return 1;
-    }
-    long hz = sysconf(_SC_CLK_TCK);
-    if (hz <= 0)
-        hz = 100;
-    printf("%.2f %.2f\n%.2f %.2f\n",
-           (double)t.tms_utime / hz,
-           (double)t.tms_stime / hz,
-           (double)t.tms_cutime / hz,
-           (double)t.tms_cstime / hz);
-    last_status = 0;
-    return 1;
-}
 
 /* Print a usage summary of the available builtin commands. */
 int builtin_help(char **args) {
@@ -808,127 +583,6 @@ int builtin_type(char **args) {
     return 1;
 }
 
-static int sig_from_name(const char *name)
-{
-    if (!name || !*name)
-        return -1;
-    if (isdigit((unsigned char)name[0]))
-        return atoi(name);
-    if (strncasecmp(name, "SIG", 3) == 0)
-        name += 3;
-    for (int i = 0; sig_map[i].n; i++) {
-        if (strcasecmp(name, sig_map[i].n) == 0)
-            return sig_map[i].v;
-    }
-    return -1;
-}
-
-static const char *name_from_sig(int sig)
-{
-    for (int i = 0; sig_map[i].n; i++) {
-        if (sig_map[i].v == sig)
-            return sig_map[i].n;
-    }
-    return NULL;
-}
-
-/* Assign commands to run when specified signals are received. */
-int builtin_trap(char **args)
-{
-    if (!args[1]) {
-        fprintf(stderr, "usage: trap [-p] [command] SIGNAL...\n");
-        return 1;
-    }
-
-    if (strcmp(args[1], "-p") == 0) {
-        if (args[2]) {
-            fprintf(stderr, "usage: trap -p\n");
-            return 1;
-        }
-        if (exit_trap_cmd)
-            printf("trap '%s' EXIT\n", exit_trap_cmd);
-        for (int s = 1; s < NSIG; s++) {
-            if (trap_cmds[s]) {
-                const char *name = name_from_sig(s);
-                if (name)
-                    printf("trap '%s' %s\n", trap_cmds[s], name);
-                else
-                    printf("trap '%s' %d\n", trap_cmds[s], s);
-            }
-        }
-        last_status = 0;
-        return 1;
-    }
-
-    char *cmd = NULL;
-    int idx = 1;
-    if (args[2]) {
-        cmd = args[1];
-        idx = 2;
-    }
-
-    for (int i = idx; args[i]; i++) {
-        if (strcasecmp(args[i], "EXIT") == 0 || strcmp(args[i], "0") == 0) {
-            free(exit_trap_cmd);
-            exit_trap_cmd = cmd ? strdup(cmd) : NULL;
-            continue;
-        }
-        int sig = sig_from_name(args[i]);
-        if (sig <= 0 || sig >= NSIG) {
-            fprintf(stderr, "trap: invalid signal %s\n", args[i]);
-            continue;
-        }
-        free(trap_cmds[sig]);
-        trap_cmds[sig] = cmd ? strdup(cmd) : NULL;
-
-        struct sigaction sa;
-        sigemptyset(&sa.sa_mask);
-        sa.sa_flags = 0;
-        sa.sa_handler = cmd ? trap_handler : SIG_DFL;
-        sigaction(sig, &sa, NULL);
-    }
-    return 1;
-}
-
-/* Signal a loop to terminate after the current iteration. */
-int builtin_break(char **args)
-{
-    int n = 1;
-    if (args[1]) {
-        char *end;
-        errno = 0;
-        long val = strtol(args[1], &end, 10);
-        if (*end != '\0' || errno != 0 || val <= 0) {
-            fprintf(stderr, "usage: break [N]\n");
-            return 1;
-        }
-        n = (int)val;
-    }
-    if (n > loop_depth)
-        n = loop_depth;
-    loop_break = n;
-    return 1;
-}
-
-/* Skip directly to the next iteration of the innermost loop. */
-int builtin_continue(char **args)
-{
-    int n = 1;
-    if (args[1]) {
-        char *end;
-        errno = 0;
-        long val = strtol(args[1], &end, 10);
-        if (*end != '\0' || errno != 0 || val <= 0) {
-            fprintf(stderr, "usage: continue [N]\n");
-            return 1;
-        }
-        n = (int)val;
-    }
-    if (n > loop_depth)
-        n = loop_depth;
-    loop_continue = n;
-    return 1;
-}
 
 /* Display or modify the file creation mask. */
 /* Helper to print a mask in symbolic form like u=rwx,g=rx,o=rx. */
